backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pillow==11.0.0
